ISO Scala jQuery - Rename package - Clean up
